% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbGetBusinessInvoices.R
\name{fbGetBusinessInvoices}
\alias{fbGetBusinessInvoices}
\title{Get Business Invoices}
\usage{
fbGetBusinessInvoices(
  business_id,
  start_date = NULL,
  end_date = NULL,
  issue_start_date = NULL,
  issue_end_date = NULL,
  invoice_id = NULL,
  type = NULL,
  api_version = getOption("rfacebookstat.api_version"),
  username = getOption("rfacebookstat.username"),
  token_path = fbTokenPath(),
  access_token = getOption("rfacebookstat.access_token")
)
}
\arguments{
\item{business_id}{ID of the Facebook Business Manager.}

\item{start_date}{Start date for the billing period (YYYY-MM-DD).}

\item{end_date}{End date for the billing period (YYYY-MM-DD).}

\item{issue_start_date}{Start date when the invoice was issued (YYYY-MM-DD).}

\item{issue_end_date}{End date when the invoice was issued (YYYY-MM-DD).}

\item{invoice_id}{Filter by a specific invoice ID.}

\item{type}{Type of the document, e.g., "INV", "PRO_FORMA", "CM", "DM".}

\item{api_version}{Current Facebook API version.}

\item{username}{your username on Facebook}

\item{token_path}{path to dir with credentials}

\item{access_token}{Your facebook API token}
}
\value{
Data frame with business invoices.
}
\description{
\code{fbGetBusinessInvoices} get invoices for a Facebook Business Manager.
}
\author{
Alexey Seleznev
}
