\name{fbGetLongTimeToken}
\alias{fbGetLongTimeToken}

\title{Get API facebook long time token.}
\description{Change your short time token to long time token.}
\usage{
fbGetLongTimeToken(client_id= NULL,client_secret = NULL,fb_exchange_token = NULL)
}

\arguments{
  \item{client_id}{ Your App ID.}
  \item{client_secret}{Your App secret.}
  \item{fb_exchange_token}{Your short time API token.}
}
\value{API token}

\author{Alexey Seleznev}

