\name{fbSetters}
\alias{fbSetters}
\alias{fbSetUsername}
\alias{fbSetAccount}
\alias{fbSetBusinessId}
\alias{fbSetTokenPath}
\alias{fbSetApiVersion}
\title{
Set rfacebookstat options
}
\description{
Set rfacebookstat options in current R session
}
\usage{
fbSetUsername(username)
fbSetAccount(accounts_ids) 
fbSetBusinessId(business_ids)
fbSetTokenPath(token_path)
fbSetApiVersion(api_version)
}
\arguments{
  \item{accounts_ids}{Facebook Ad Account ID.}
  \item{business_ids}{IDs by your Business Manager}
  \item{username}{your username on Facebook}
  \item{token_path}{path to dir with credentials}
  \item{api_version}{Current Facebook API version.}
}
\author{
Alexey Seleznev
}
